+++ tests/check_menu_pages.gd
@@ -54,7 +54,7 @@ func check() -> void:
 	var controls = scene.get_node("Controls")
 	var controller_event := InputEventJoypadButton.new()
 	controller_event.device = 3
-	controller_event.button_index = JOY_BUTTON_RIGHT_STICK
+	controller_event.button_index = JOY_BUTTON_LEFT_STICK
 	controller_event.pressed = true
 	root.push_input(controller_event)
 	panel.next_page.grab_focus()